ops-agent-exception/requirements.txt
# 异常分析框架依赖

# 配置和工具
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.0.0
rich>=13.0.0

# 日志
# 日志库根据实际需求添加

# 数据源连接
# 根据实际数据源添加相应的客户端库

# LLM 支持（可选）
# openai>=1.0.0
# langchain>=0.1.0

# 统计和数据分析（可选）
# numpy>=1.24.0
# pandas>=2.0.0
# scipy>=1.10.0

